Refactor and enhance code documentation, with task package restructuring.
Reorganized background task-related classes into a new `task.tasks` package.
Added comprehensive JavaDoc comments across plugin classes for better maintainability.
Upgraded `akz-service` dependency to version 3.0.0.
Improved exception messages and synchronized methods within utility and handler classes.
