+++ plugin/src/main/java/org/akazukin/intellij/background/actions/SetBackgroundAction.java
@@ -3,25 +3,29 @@ package org.akazukin.intellij.background.actions;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import org.akazukin.intellij.background.PluginHandler;
-import org.akazukin.intellij.background.settings.Config;
-import org.akazukin.intellij.background.task.SetRandomBackgroundTask;
+import org.akazukin.intellij.background.task.tasks.SetRandomBackgroundTask;
 import org.jetbrains.annotations.NotNull;
 
+/**
+ * An action to set a random background image from cached images in the plugin.
+ * <p>
+ * This action uses the task manager from the plugin to execute the
+ * {@link SetRandomBackgroundTask} responsible for selecting and applying
+ * a random background image based on cached images.
+ * <p>
+ * Use this action to dynamically change the background image within the editor
+ * or frame environments provided by the plugin. It is designed to update the
+ * visuals based on user-configured and validated image caches.
+ */
 public final class SetBackgroundAction extends AnAction {
-
     public SetBackgroundAction() {
         super("Set Random Background Image",
             "Set random background image from cached images", null);
     }
 
     @Override
-    public synchronized void actionPerformed(@NotNull final AnActionEvent e) {
-        final Config.State state = Config.getInstance();
-        if (state.isAutoChangeEnabled()) {
-            PluginHandler.getPlugin().getScheduler().schedule();
-        } else {
-            PluginHandler.getPlugin().getTaskMgr()
-                .getServiceByImplementation(SetRandomBackgroundTask.class).get();
-        }
+    public void actionPerformed(@NotNull final AnActionEvent e) {
+        PluginHandler.getPlugin().getTaskMgr()
+            .getServiceByImplementation(SetRandomBackgroundTask.class).get();
     }
 }
