--- plugin/src/main/java/org/akazukin/intellij/background/task/ITask.java
@@ -1,9 +1,0 @@
-package org.akazukin.intellij.background.task;
-
-import org.akazukin.service.IService;
-
-import java.util.function.Supplier;
-
-public interface ITask<T> extends Supplier<T>, IService {
-    String getTaskName();
-}
