+++ plugin/src/main/java/org/akazukin/intellij/background/task/tasks/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides some tasks for the background task.
+ */
+package org.akazukin.intellij.background.task.tasks;
